requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
setuptools==78.1.0
Werkzeug==3.1.3
wheel==0.45.1
